requirements.txt
requests==2.31.0
gspread==5.11.1
oauth2client==4.1.3
psycopg2==2.9.6
python-telegram-bot==13.15
schedule==1.2.0
